pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "beautiful-photo"
version = "0.0.1"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "scipy>=1.16.3",
]

[tool.setuptools.packages.find]
where = ["src"]
